AST, Type: Move AST nodes concerning types to ast module. Simplify class hierarchy.
