+++ compiler/parse.py
@@ -17,7 +17,7 @@ from compiler import ast, lex, type
 
 
 def _track(p):
-    if isinstance(p[1], (ast.Node, type.Type)):
+    if isinstance(p[1], ast.Node):
         if p[0] is not p[1]:
             p[0].copy_pos(p[1])
     else:
@@ -124,7 +124,7 @@ class Parser:
                         | FLOAT
                         | INT
                         | UNIT"""
-        p[0] = type.builtin_map[p[1]]()
+        p[0] = ast.builtin_map[p[1]]()
         _track(p)
 
     def p_derived_type(self, p):
@@ -139,9 +139,9 @@ class Parser:
         """array_type : ARRAY LBRACKET star_comma_seq RBRACKET OF type
                       | ARRAY OF type"""
         if len(p) == 7:
-            p[0] = type.Array(p[6], p[3])
+            p[0] = ast.Array(p[6], p[3])
         else:
-            p[0] = type.Array(p[3])
+            p[0] = ast.Array(p[3])
         _track(p)
 
     def p_star_comma_seq(self, p):
@@ -155,17 +155,17 @@ class Parser:
 
     def p_function_type(self, p):
         """function_type : type ARROW type"""
-        p[0] = type.Function(p[1], p[3])
+        p[0] = ast.Function(p[1], p[3])
         _track(p)
 
     def p_ref_type(self, p):
         """ref_type : type REF"""
-        p[0] = type.Ref(p[1])
+        p[0] = ast.Ref(p[1])
         _track(p)
 
     def p_user_type(self, p):
         """user_type : GENID"""
-        p[0] = type.User(p[1])
+        p[0] = ast.User(p[1])
         _track(p)
 
     def p_empty(self, p):
@@ -261,12 +261,12 @@ class Parser:
     def p_bconst_simple_expr(self, p):
         """bconst_simple_expr : TRUE
                               | FALSE"""
-        p[0] = ast.ConstExpression(type.Bool(), p[1])
+        p[0] = ast.ConstExpression(ast.Bool(), p[1])
         _track(p)
 
     def p_cconst_simple_expr(self, p):
         """cconst_simple_expr : CCONST"""
-        p[0] = ast.ConstExpression(type.Char(), p[1])
+        p[0] = ast.ConstExpression(ast.Char(), p[1])
         _track(p)
 
     def p_conid_simple_expr(self, p):
@@ -276,12 +276,12 @@ class Parser:
 
     def p_iconst_simple_expr(self, p):
         """iconst_simple_expr : ICONST"""
-        p[0] = ast.ConstExpression(type.Int(), p[1])
+        p[0] = ast.ConstExpression(ast.Int(), p[1])
         _track(p)
 
     def p_fconst_simple_expr(self, p):
         """fconst_simple_expr : FCONST"""
-        p[0] = ast.ConstExpression(type.Float(), p[1])
+        p[0] = ast.ConstExpression(ast.Float(), p[1])
         _track(p)
 
     def p_genid_simple_expr(self, p):
@@ -291,12 +291,12 @@ class Parser:
 
     def p_sconst_simple_expr(self, p):
         """sconst_simple_expr : SCONST"""
-        p[0] = ast.ConstExpression(type.String(), p[1])
+        p[0] = ast.ConstExpression(ast.String(), p[1])
         _track(p)
 
     def p_uconst_simple_expr(self, p):
         """uconst_simple_expr : LPAREN RPAREN"""
-        p[0] = ast.ConstExpression(type.Unit())
+        p[0] = ast.ConstExpression(ast.Unit())
         _track(p)
 
     def p_delete_expr(self, p):
@@ -390,26 +390,26 @@ class Parser:
     def p_bconst_simple_pattern(self, p):
         """bconst_simple_pattern : TRUE
                                  | FALSE"""
-        p[0] = ast.ConstExpression(type.Bool(), p[1])
+        p[0] = ast.ConstExpression(ast.Bool(), p[1])
         _track(p)
 
     def p_cconst_simple_pattern(self, p):
         """cconst_simple_pattern : CCONST"""
-        p[0] = ast.ConstExpression(type.Char(), p[1])
+        p[0] = ast.ConstExpression(ast.Char(), p[1])
         _track(p)
 
     def p_fconst_simple_pattern(self, p):
         """fconst_simple_pattern : FPLUS FCONST
                                  | FCONST"""
         if len(p) == 3:
-            p[0] = ast.ConstExpression(type.Float(), p[2])
+            p[0] = ast.ConstExpression(ast.Float(), p[2])
         else:
-            p[0] = ast.ConstExpression(type.Float(), p[1])
+            p[0] = ast.ConstExpression(ast.Float(), p[1])
         _track(p)
 
     def p_mfconst_simple_pattern(self, p):
         """mfconst_simple_pattern : FMINUS FCONST"""
-        p[0] = ast.ConstExpression(type.Float(), -p[2])
+        p[0] = ast.ConstExpression(ast.Float(), -p[2])
         _track(p)
 
     def p_genid_simple_pattern(self, p):
@@ -421,14 +421,14 @@ class Parser:
         """iconst_simple_pattern : PLUS ICONST
                                  | ICONST"""
         if len(p) == 3:
-            p[0] = ast.ConstExpression(type.Int(), p[2])
+            p[0] = ast.ConstExpression(ast.Int(), p[2])
         else:
-            p[0] = ast.ConstExpression(type.Int(), p[1])
+            p[0] = ast.ConstExpression(ast.Int(), p[1])
         _track(p)
 
     def p_miconst_simple_pattern(self, p):
         """miconst_simple_pattern : MINUS ICONST"""
-        p[0] = ast.ConstExpression(type.Int(), -p[2])
+        p[0] = ast.ConstExpression(ast.Int(), -p[2])
         _track(p)
 
     def p_new_expr(self, p):
